LG-340 Use Twilio Verify for international SMS
**Why**: It purports to be more reliable than the Twilio REST API.
Another benefit is that they provide localized messages in a bunch of
languages (whereas we only support English, French, and Spanish).
By default, the Verify service uses their own generated code, but they
also allow you to use your own code and validation, which is what I've
done in this PR.

The advantage of using Twilio's code is that it can be cheaper because
once a user requests an SMS, it is valid for 10 minutes, and if the user
requests another SMS while the first one is still valid, Twilio will
resend the original code at no additional cost.

The disadvantage is that we no longer retain control of how the code
is generated, and how it is validated. It would also require a second
API call to verify the code entered by the user.

From the documentation, it seems that if we want to take advantage of
the localization, we cannot use our own custom text. We are meeting with
Twilio to get confirmation.